+++ app/lib/services/local_storage_service.dart
@@ -1,6 +1,5 @@
 import 'dart:io';
 
-import 'package:insanichess/insanichess.dart' as insanichess;
 import 'package:insanichess_sdk/insanichess_sdk.dart';
 import 'package:localstorage/localstorage.dart';
 import 'package:path_provider/path_provider.dart';
@@ -25,8 +24,14 @@ class LocalStorageService {
   static const String _settingsFile = 'settings.json';
 
   Future<void> saveGame(InsanichessGame game) async {
+    final String directoryPath = await _gamesPath;
+    final Directory directory = Directory(directoryPath);
+    if (!await directory.exists()) {
+      directory.create();
+    }
+
     final File f = File(
-      '${await _gamesPath}${Platform.pathSeparator}${game.id}',
+      '$directoryPath${Platform.pathSeparator}${game.id}',
     );
 
     await f.writeAsString(game.toICString(), flush: true);
@@ -37,32 +42,7 @@ class LocalStorageService {
       '${await _gamesPath}${Platform.pathSeparator}$id',
     );
 
-    final List<String> lines = await f.readAsLines();
-
-    final List<String> timeControlString = lines[2].split(' ');
-    final List<String> remainingTimeString = lines[3].split(' ');
-
-    final List<insanichess.PlayedMove> moves = <insanichess.PlayedMove>[];
-    // for (int i = 5; i < lines.length; i++) {
-    //   final List<String> splittedLine = lines[i].split(' ');
-    //   final insanichess.PlayedMove whiteMove = insanichess.PlayedMove(insanichess.Square(), insanichess.Square());
-    // }
-
-    return InsanichessGame.fromPosition(
-      id: lines[0],
-      whitePlayer: const InsanichessPlayer.testWhite(),
-      blackPlayer: const InsanichessPlayer.testBlack(),
-      timeControl: InsanichessTimeControl(
-        initialTime: Duration(seconds: int.parse(timeControlString[0])),
-        incrementPerMove: Duration(seconds: int.parse(timeControlString[1])),
-      ),
-      position: positionFromFen(lines[4]),
-      remainingTimeWhite:
-          Duration(milliseconds: int.parse(remainingTimeString[0])),
-      remainingTimeBlack:
-          Duration(milliseconds: int.parse(remainingTimeString[1])),
-      gameHistory: insanichess.GameHistory.withMoves(moves: moves),
-    );
+    return InsanichessGame.fromICString(await f.readAsString());
   }
 
   Future<void> saveSettings(InsanichessSettings settings) async {